bug fix on BSON serialization(BsonConverter)

+++ src/main/java/com/doodream/rmovjs/client/RMIClient.java
@@ -378,7 +378,7 @@ public class RMIClient implements InvocationHandler, Comparable<RMIClient>  {
 
     @Override
     public int compareTo(@NotNull RMIClient o) {
-        return Math.toIntExact(getMeasuredPing() - o.getMeasuredPing());
+        return (int) (getMeasuredPing() - o.getMeasuredPing());
     }
 
     private long getMeasuredPing() {
